chore: Preparing SqlDataTable for veterinary specific needs
The code changes include adding a new sub icon to the SQL data table component. This icon allows users to add a new sub entry for a specific data element. The functionality is triggered by clicking on the icon. This enhancement improves the usability of the SQL data table component by providing an intuitive way to add new sub entries.