chore: Update MongoDB connection URI for production environment
This commit updates the MongoDB connection URI in the `database.js` file to include additional options for the production environment. The new URI includes the `retryWrites=true` and `w=majority` options, as well as the `appName=arcadia-zoo` parameter. This change ensures a more reliable and optimized connection to the MongoDB database in the production environment.